# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Constituent Touchpoint Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      parameters:
        - { name: mine, in: query, schema: { type: boolean } }
        - { name: status, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CampaignSummary" }
    post:
      operationId: createCampaign
      tags: [campaigns]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
  /campaigns/{id}/archive:
    post:
      operationId: archiveCampaign
      tags: [campaigns]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
  /campaigns/{id}/void:
    post:
      operationId: voidCampaign
      tags: [campaigns]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }

  # ───────── Audience
  /campaigns/{id}/audience:
    post:
      operationId: uploadAudience
      tags: [audience]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AudienceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AudienceUploadResult" }
    get:
      operationId: getAudienceSummary
      tags: [audience]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AudienceUploadResult" }

  # ───────── Touches
...
  /donors/{donorId}/touchpoints:
    get:
      operationId: getDonorTouchpoints
      tags: [donors]
      parameters:
        - { name: donorId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DonorLookup" }

  # ───────── Reports
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    SessionUser:
      type: object
      required: [id, email, name, role, active, piiAcknowledged]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [standard, admin, super_admin] }
        active: { type: boolean }
        piiAcknowledged: { type: boolean }
...
    CampaignType:
      type: object
      required: [id, name, active, systemDefault]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        active: { type: boolean }
        systemDefault: { type: boolean }
...
    CampaignInput:
      type: object
      required: [name, campaignTypeIds]
      properties:
        name: { type: string }
        owningUnit: { type: string }
        intendedSendStartDate: { type: ["string", "null"], format: date }
        audienceDescription: { type: string }
        salesforceCampaignId: { type: string }
        internalNotes: { type: string }
        campaignTypeIds:
          type: array
          items: { type: integer }
    CampaignUpdate:
      type: object
      properties:
        name: { type: string }
        owningUnit: { type: string }
        intendedSendStartDate: { type: ["string", "null"], format: date }
        audienceDescription: { type: string }
        salesforceCampaignId: { type: string }
        internalNotes: { type: string }
        campaignTypeIds:
          type: array
          items: { type: integer }
    CampaignSummary:
      type: object
      required: [id, name, status, submittedByName, owningUnit, createdAt, touchCount, audienceSize, campaignTypes]
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string }
        submittedByName: { type: string }
        submittedByUserId: { type: integer }
        owningUnit: { type: ["string", "null"] }
        intendedSendStartDate: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
        exportedAt: { type: ["string", "null"], format: date-time }
        touchCount: { type: integer }
        audienceSize: { type: integer }
        campaignTypes:
          type: array
          items: { type: string }
    Campaign:
      type: object
      required: [id, name, status, submittedByUserId, submittedByName, createdAt, campaignTypes]
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string }
        owningUnit: { type: ["string", "null"] }
        submittedByUserId: { type: integer }
        submittedByName: { type: string }
        intendedSendStartDate: { type: ["string", "null"], format: date }
        audienceDescription: { type: ["string", "null"] }
        salesforceCampaignId: { type: ["string", "null"] }
        internalNotes: { type: ["string", "null"] }
        originalRowCount: { type: integer }
        blankRowCount: { type: integer }
        validIdCount: { type: integer }
        uniqueIdCount: { type: integer }
        duplicateIdCount: { type: integer }
        rejectedIdCount: { type: integer }
        extraColumnsIgnored: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        exportedAt: { type: ["string", "null"], format: date-time }
        archivedAt: { type: ["string", "null"], format: date-time }
        voidedAt: { type: ["string", "null"], format: date-time }
        campaignTypes:
          type: array
          items: { $ref: "#/components/schemas/CampaignType" }
    AudienceInput:
      type: object
      properties:
        # Provide one of: rawText, googleSheetUrl
        rawText: { type: string }
        googleSheetUrl: { type: string }
        hasHeader: { type: boolean }
        columnIndex: { type: integer }
    AudienceUploadResult:
      type: object
      required: [originalRowCount, blankRowCount, validCount, uniqueCount, duplicateCount, rejectedCount, extraColumnsIgnored, rejectedSamples, duplicateSamples]
      properties:
        originalRowCount: { type: integer }
        blankRowCount: { type: integer }
        validCount: { type: integer }
        uniqueCount: { type: integer }
        duplicateCount: { type: integer }
        rejectedCount: { type: integer }
        extraColumnsIgnored: { type: boolean }
        detectedColumns:
          type: array
          items: { type: string }
        rejectedSamples:
          type: array
          items: { type: string }
        duplicateSamples:
          type: array
          items: { type: string }
...
    DonorLookup:
      type: object
      required: [donorId, touchpoints]
      properties:
        donorId: { type: string }
        touchpoints:
          type: array
          items:
            type: object
            required: [campaignId, campaignName, campaignStatus, channelLabel, campaignTypeLabel, sendDate, countsTowardThreshold]
            properties:
              campaignId: { type: integer }
              campaignName: { type: string }
              campaignStatus: { type: string }
              channelLabel: { type: string }
              campaignTypeLabel: { type: string }
              sendDate: { type: string, format: date }
              countsTowardThreshold: { type: boolean }
...